Remove redundancy in the gfx1030 changes
Three duplications introduced by this branch, no behavior change.

- tests/unit/test_api.py defined a _FakeBackend that was a near-copy of
  test_cli.py's FakeBackend: same five methods, same _available /
  _unsupported_metrics / _aggregated / profile_calls. Promoted the original
  to conftest.py with an `arch` parameter and pointed both files at it, so
  the two GPU-free suites share one double instead of drifting apart.

- _no_output_message was a @staticmethod that never touched self, sitting
  next to the module-level _extract_missing_counters it delegates to. Made
  it a module function alongside its helper.

- Three tests each re-imported _extract_missing_counters inside the method
  body; hoisted to the existing module-level import.

unit 351 passed / 33 skipped, integration 20 passed / 29 skipped, ruff clean.
